move to use routes.js file, logger util just for node context.
