+++ src/applications/_mock-form-ae-design-patterns/mocks/server.js
@@ -48,7 +48,7 @@ const responses = {
           generateFeatureToggles({
             aedpVADX: true,
             coeAccess: true,
-            profileUseExperimental: true,
+            profileUseExperimental: false,
           }),
         ),
       secondsOfDelay,
